fixed tunnel/entities <--> tunnel/intercept codependency
